# docs/swagger.yaml
definitions:
  gin.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/model.UserResponse'
      token:
        type: string
    type: object
  model.CategoryCreationType:
    properties:
      category_name_eng:
        type: string
      category_name_vn:
        type: string
      dph:
        type: integer
      image_url:
        type: string
    required:
    - category_name_eng
    - category_name_vn
    - dph
    type: object
  model.CategoryType:
    properties:
      category_id:
        type: string
      category_name_eng:
        type: string
      category_name_vn:
        type: string
      dph:
        type: integer
      image_url:
        type: string
    required:
    - category_id
    - category_name_eng
    - category_name_vn
    - dph
    type: object
  model.CurrencyCreationType:
    properties:
      currency_code:
        type: string
      currency_name:
        type: string
      exchange_rate:
        type: number
    required:
    - currency_code
    - currency_name
    type: object
  model.LoginInput:
    properties:
      password:
        type: string
      user_name:
        type: string
    required:
    - password
    - user_name
    type: object
  model.ProductCreationType:
    properties:
      category_id:
        type: string
      description:
        type: string
      dph:
        type: integer
      image_url:
        type: string
      product_code:
        type: string
      product_name_eng:
        type: string
      product_name_vn:
        type: string
    required:
    - category_id
    - product_code
    - product_name_eng
    - product_name_vn
    type: object
  model.SupplierCreationType:
    properties:
      currencyID:
        type: string
      description:
        type: string
      supplier_name:
        type: string
      tax_id:
        type: string
    required:
    - supplier_name
    - tax_id
    type: object
  model.UserCreationType:
    properties:
      email:
        type: string
      password:
        type: string
      user_name:
        type: string
    required:
    - email
    - password
    - user_name
    type: object
  model.UserResponse:
    properties:
      email:
        type: string
      status:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  model.UserType:
    properties:
      email:
        type: string
      status:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
info:
  contact: {}
  description: Đây là tài liệu API cho ứng dụng Hệ Thống Veggie
  title: Veggie API
  version: "1.0"
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Đăng Nhập tài khoản mới bằng email
      parameters:
      - description: User data
        in: body
        name: users
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Đăng Nhập Tài Khoản Thành Công
          schema:
            $ref: '#/definitions/gin.LoginResponse'
      summary: Đăng Nhập Tài Khoản
      tags:
      - Authorization
  /v1/categories:
    get:
      consumes:
      - application/json
      description: Tìm Nhóm Sản Phẩm
      produces:
      - application/json
      responses:
        "200":
          description: Tìm Nhóm Sản Phẩm Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tìm Nhóm Sản Phẩm
      tags:
      - Nhóm Sản Phẩm
    post:
      consumes:
      - application/json
      description: Tạo Nhóm Sản Phẩm
      parameters:
      - description: category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryCreationType'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo Nhóm Sản Phẩm Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tạo Nhóm Sản Phẩm
      tags:
      - Nhóm Sản Phẩm
    put:
      consumes:
      - application/json
      description: Cập Nhật Nhóm Sản Phẩm
      parameters:
      - description: category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/model.CategoryType'
      produces:
      - application/json
      responses:
        "200":
          description: Cập Nhật Nhóm Sản Phẩm Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Cập Nhật Nhóm Sản Phẩm
      tags:
      - Nhóm Sản Phẩm
  /v1/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Xóa Nhóm Sản Phẩm Theo Id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Xóa Nhóm Sản Phẩm Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Xóa Nhóm Sản Phẩm Theo Id
      tags:
      - Nhóm Sản Phẩm
    put:
      consumes:
      - application/json
      description: Tìm Nhóm Sản Phẩm Theo Id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tìm Nhóm Sản Phẩm Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tìm Nhóm Sản Phẩm Theo Id
      tags:
      - Nhóm Sản Phẩm
  /v1/currencies:
    get:
      consumes:
      - application/json
      description: Tìm Loại Tiền Tệ
      produces:
      - application/json
      responses:
        "200":
          description: Tìm Loại Tiền Tệ Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tìm Loại Tiền Tệ
      tags:
      - Tiền Tệ
    post:
      consumes:
      - application/json
      description: Tạo Tiền Tệ mới
      parameters:
      - description: Curency data
        in: body
        name: currency
        required: true
        schema:
          $ref: '#/definitions/model.CurrencyCreationType'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo Tiền Tệ Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tạo Tiền Tệ
      tags:
      - Tiền Tệ
  /v1/currencies/{id}:
    delete:
      consumes:
      - application/json
      description: Xóa Tiền Tệ Theo ID
      parameters:
      - description: Curency id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Xóa Tiền Tệ Theo ID Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Xóa Tiền Tệ Theo ID
      tags:
      - Tiền Tệ
    put:
      consumes:
      - application/json
      description: Tìm Tiền Tệ Theo ID
      parameters:
      - description: Curency id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tìm Tiền Tệ Theo ID Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tìm Tiền Tệ Theo ID
      tags:
      - Tiền Tệ
  /v1/products:
    post:
      consumes:
      - application/json
      description: Tạo Sản Phẩm Mới
      parameters:
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/model.ProductCreationType'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo Tài Khoản Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tạo Sản Phẩm
      tags:
      - Sản Phẩm
  /v1/supplier:
    post:
      consumes:
      - application/json
      description: Tạo Nhà Cung Cấp
      parameters:
      - description: supplier
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/model.SupplierCreationType'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo Nhà Cung Cấp Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Tạo Nhà Cung Cấp
      tags:
      - Nhà Cung Cấp
  /v1/users:
    get:
      consumes:
      - application/json
      description: Lấy danh sách tài khoản
      produces:
      - application/json
      responses:
        "200":
          description: Lấy danh sách tài khoản Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Lấy danh sách tài khoản
      tags:
      - Tài Khoản
    post:
      consumes:
      - application/json
      description: Tạo tài khoản mới bằng email
      parameters:
      - description: User data
        in: body
        name: users
        required: true
        schema:
          $ref: '#/definitions/model.UserCreationType'
      produces:
      - application/json
      responses:
        "200":
          description: Tạo Tài Khoản Thành Công
          schema:
            type: object
      summary: Tạo Tài Khoản
      tags:
      - Authorization
    put:
      consumes:
      - application/json
      description: Cập nhật tài khoản
      parameters:
      - description: User update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserType'
      produces:
      - application/json
      responses:
        "200":
          description: Cập nhật tài khoản Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Cập nhật tài khoản
      tags:
      - Tài Khoản
  /v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Xóa tài khoản theo id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Xóa tài khoản theo id Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Xóa tài khoản theo id
      tags:
      - Tài Khoản
    put:
      consumes:
      - application/json
      description: Lấy tài khoản theo id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lấy tài khoản theo id Thành Công
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Lấy tài khoản theo id
      tags:
      - Tài Khoản
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"